scraper/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
fake-useragent==1.5.1
anthropic==0.29.0
pydantic==2.8.2
python-dotenv==1.0.1
pymongo==4.8.0
aiohttp==3.9.5
asyncio==3.4.3
urllib3==2.2.2
certifi==2024.7.4
charset-normalizer==3.3.2
cloudscraper==1.2.71
httpx==0.27.0 